pyproject.toml
[tool.poetry]
name = "dags"
version = "0.1.0"
description = ""
authors = ["Pabloo22 <pablete.arino@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
apache-airflow = "2.8.2"
apache-airflow-providers-apache-spark = "4.7.1"
pyspark = "3.4.0"
ipykernel = "^6.29.3"
pandas = "^2.2.1"
pandas-stubs = "^2.2.1.240316"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.0"
types-requests = "^2.31.0.20240403"
types-beautifulsoup4 = "^4.12.0.20240229"
folium = "^0.16.0"
scipy = "^1.13.0"
branca = "^0.7.2"
streamlit-folium = "^0.20.0"
psycopg2-binary = "^2.9.9"
plotly = "^5.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79